Switched to "mattn" SQLite3 driver package
This driver uses the native SQLite so it's faster. It's also more mature, with
features the modernc one lacks.

It turns out I was misusing SQL `$` parameters: in SQLite at least these are
named not ordinal. For some reason the modernc driver didn't mind, but the
mattn one was getting args in the wrong order until I switched to the correct
`?n` syntax.