Fixed deadlock with QueryResultIterator
In-memory mode only has a single connection, so it's possible to
deadlock if the client has a QueryResultIterator (which owns a
sql.Rows, which uses a db connection) and tries to make another
call.
Worked around this by checking for in-memory mode, and in that mode
running the entire query in advance and having the iterator read the
rows out of an array. Less efficient but it works.
Need to decide whether this deadlock can happen in persistent mode
where many connections are available...